apps/requirements.txt
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.6.1
exceptiongroup==1.2.2
flatten-json==0.1.14
h11==0.14.0
idna==3.10
nslookup==1.8.1
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing-extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
